Update#14 Tests

+++ src/main/java/dto/response/UserResponseDto.java
@@ -16,5 +16,5 @@ public record UserResponseDto(
 
         Role role,
 
-        Integer noteCount
+        Long noteCount
 ) {}
